refactor PersonCard
